합치기 완료
