+++ castlevania/20180820/battleScene.h
@@ -6,15 +6,8 @@ class enemyManager;
 class camel;
 class player;
 //·ë
-class roomManager;
-class hallwayRoom1;
-class hallwayRoom2;
-class hallwayRoom3;
-class FountainRoom;
-class gateRoom;
-class SaveRoom;
-class CastleHallway;
-class BrickStaircaseroom;
+
+
 
 
 
@@ -31,15 +24,7 @@ private:
 	zombi* m_zombi;
 
 	//·Ò
-	hallwayRoom1*	m_phallwayRoom1;
-	hallwayRoom2*	m_phallwayRoom2;
-	hallwayRoom3*	m_phallwayRoom3;
-	FountainRoom*	m_pFountainRoom;
-	gateRoom*		m_gateroom;
-	SaveRoom*		m_SaveRoom;
-	roomManager*	m_RoomMgr;
-	CastleHallway*  m_CastleHallway;
-	BrickStaircaseroom* m_BrickStaircaseroom;
+
 
 	player*			m_pPlayer;
