+++ castlevania/20180820/mainGame.cpp
@@ -8,6 +8,16 @@
 #include "menuScene.h"
 #include "player.h"
 
+//·ë
+#include "hallwayRoom1.h"
+#include "hallwayRoom2.h"
+#include "hallwayRoom3.h"
+#include "FountainRoom.h"
+#include"gateroom.h"
+#include "SaveRoom.h"
+#include "CastleHallway.h"
+#include "BrickStaircaseroom.h"
+
 
 /*
 	°ÔÀÓ window Ã¢¶ç¿ì±â (winmain () )
@@ -42,6 +52,33 @@ HRESULT mainGame::init()
 
 	EFFECTMANAGER->addEffect("explosion1", "image/explosion.bmp", 832, 62, 32, 62, 30, 10);
 
+	// ·ë ÀÌ´Ï¼È¶óÀÌÁî
+	m_phallwayRoom1 = new hallwayRoom1;
+	ROOMMANAGER->addRoom("hallwayRoom1", m_phallwayRoom1);
+
+	m_phallwayRoom2 = new hallwayRoom2;
+	ROOMMANAGER->addRoom("hallwayRoom2", m_phallwayRoom2);
+
+	m_phallwayRoom3 = new hallwayRoom3;
+	ROOMMANAGER->addRoom("hallwayRoom3", m_phallwayRoom3);
+
+	m_pFountainRoom = new FountainRoom;
+	ROOMMANAGER->addRoom("FountainRoom", m_pFountainRoom);
+
+	m_gateroom = new gateRoom;
+	ROOMMANAGER->addRoom("gateroom", m_gateroom);
+
+	m_SaveRoom = new SaveRoom;
+	ROOMMANAGER->addRoom("saveroom", m_SaveRoom);
+
+	m_CastleHallway = new CastleHallway;
+	ROOMMANAGER->addRoom("CastleHallwayy", m_CastleHallway);
+
+
+	m_BrickStaircaseroom = new BrickStaircaseroom;
+	ROOMMANAGER->addRoom("BrickStaircaseroom", m_BrickStaircaseroom);
+
+	ROOMMANAGER->changeRoom("hallwayRoom1");
 
 
 	/*m_pTitleScene = new titleScene;
