+++ castlevania/20180820/mainGame.h
@@ -7,6 +7,18 @@ class menuScene;
 class spaceShip;
 class player;
 
+class roomManager;
+class hallwayRoom1;
+class hallwayRoom2;
+class hallwayRoom3;
+class FountainRoom;
+class gateRoom;
+class SaveRoom;
+class CastleHallway;
+class BrickStaircaseroom;
+
+
+
 class mainGame
 {
 private:
@@ -33,6 +45,16 @@ private:
 	scoreScene*		m_pscoreScene;
 	selectScene*	m_pselectScene;
 	menuScene*		m_pmenuScene;
+
+	hallwayRoom1*	m_phallwayRoom1;
+	hallwayRoom2*	m_phallwayRoom2;
+	hallwayRoom3*	m_phallwayRoom3;
+	FountainRoom*	m_pFountainRoom;
+	gateRoom*		m_gateroom;
+	SaveRoom*		m_SaveRoom;
+	roomManager*	m_RoomMgr;
+	CastleHallway*  m_CastleHallway;
+	BrickStaircaseroom* m_BrickStaircaseroom;
 
 public:
 	HRESULT init();
